Split JobViewController Up [skip ci]
