+++ Wemotely/Controllers/JobViewController+Segues.swift
@@ -1,0 +1,15 @@
+import Foundation
+
+extension JobViewController {
+    func setupSegue(job: Job) {
+        jobRecord = job
+
+        if let split = splitViewController {
+            navigationItem.leftItemsSupplementBackButton = true
+            navigationItem.leftBarButtonItem = split.displayModeButtonItem
+        }
+
+        navigationController?.navigationBar.prefersLargeTitles = false
+        navigationItem.largeTitleDisplayMode = .never
+    }
+}
