+++ Wemotely/Support/Styles/job.css
@@ -1,0 +1,19 @@
+#container {
+    margin: 20px;
+    font-family: Arial, Helvetica, sans-serif;
+    font-size: 40px;
+    line-height: 1.5;
+}
+
+#container img {
+    display: block;
+    margin-left: auto;
+    margin-right: auto;
+    margin-bottom: 30px;
+    width: 50%;
+}
+
+#container a, #container a:visited {
+    color: #39c;
+    text-decoration: none;
+}
